Move printf_formatter to goto-programs
The only users of this code are in goto-programs, and printf_formatter has no
dependency on ansi-c code.
